+++ src/widgets/progressbar.cpp
@@ -149,7 +149,7 @@ namespace gcn
         {
             // Infinite scrollbar
             progressWidth = getWidth() / 5;
-            int barX = getWidth() * mValue / 100;
+            int barX = getWidth() * static_cast<int>(mValue) / 100;
 
             if (barX + progressWidth > getWidth())
             {
@@ -164,7 +164,7 @@ namespace gcn
         else
         {
             // Standard scrollbar
-            progressWidth = getWidth() * mValue / (mEnd - mStart);
+            progressWidth = getWidth() * static_cast<int>(mValue) / static_cast<int>(mEnd - mStart);
             graphics->fillRectangle(Rectangle(0, 0, progressWidth, getHeight()));
         }
